Version 4
FoodRepository.kt
FoodRepositoryImpl.kt
GetUsesCase
