+++ app/src/main/java/com/example/food/domain/repository/FoodRepository.kt
@@ -1,0 +1,8 @@
+package com.example.food.domain.repository
+
+import com.example.food.domain.model.FoodItem
+
+interface FoodRepository {
+
+    suspend fun getItem(name : String) : ArrayList<FoodItem>
+}
