--- app/src/main/java/com/example/food/presentation/model/Category.kt
@@ -1,5 +1,0 @@
-package com.example.food.presentation.model
-
-data class Category(
-    val category: List<FoodItem>
-)
